Support unpickling of images without their source.
A user may load images, analyse them, and save the object for later use. THe later usage may be in the absence of the sources (other computer, for example). The Images object is then only necessary for the attributes of the image, since the images themselves have already been analyzed. This commit allows to load (unpickle) images objects even in the absence of the source, to fit this need.