Fix small mistakes
In Parameters.__getattr__, looking for the attr in the defaults dict before calling it was unnecessary and would prevent the default dict to look into its own defaults.
In Dataset.test_detection_conditions, the transient Dataset was initialized with a path, which would create a new Images object. Initializing it with the existing Images object for gain of memory and speed.