Create new Experiment object.
New Experiment design fitting with the new Images design.
Experiments now inherit from Parameters.
Sources and datasets are now completely separated. A source can be used as a dataset or for several datasets. Experiments can have several sources and several datasets. All datasets' sources are located in the Experiment's source.
Datasets now use Experiment as their _defaults dict. This means that any value not in the Dataset will be fetched in the Experiment. This is more elegant than forcing values into Dataset everytime a value is set in the Experiment.