Fix import hooks.
CATS can now take detection, linkage, filtration and analysis functions as individual files. This is useful for extensibility and for readability.
Each category is a module. Its files are deposited in the respective directories, each file containing one/several functions of interest indicated by the __all__ array of the file. These files are considered to be modules by Python.
Each category's __init__ file will load the files in its directory that do not start with '_' and end with '.py' as modules. It will extract the functions indicated in the '__all__' list of the file into its own globals. It will also put these functions names in its own '__all__'.
This means that all files' defined functions in the categories (which are subpackages, to Python's eye) are loaded as the subpackage's function. Hence, the subpackage is a module.

For example, a directory structure as is:
CATS/analysis/plot.py (__all__ = ['plot'])
              survival.py (__all__ = ['survival', 'lifetime'])
              fun_test.py (__all__ = ['fun_test'])
>>> import CATS.analysis
>>> dir(CATS.analysis)
'plot',
'survival',
'lifetime',
'fun_test'
[...]

Analysis functions are directly added to the Experiment class. All functions in detection, linkage, filtration and analysis should be written as methods.