Replace the analysis module with extensions.
The analysis module was specific to the Experiment object and was only supposed to supported analysis methods. However, every object in CATS may need to be extended, including Experiment, with methods that are not related to analysis of data. Hence, the extensions, while supporting analysis functions are more versatile and hence better suited.