Implement absolute limits for ROIs.
When making a ROI of a ROI, the .x/.y/.t limits are relative to the source object, which has its own limits, not the source images. abs_x/abs_y/abs_t calculate the limits from the source images, making it easier to trace back the ROI.
Also, this makes __repr__() more exact.