Implement a better linking algo for stationary particles
The old linking algorithm was erratic and would allow movement > max_disp. Additionally, it would loose tracks and merge wrong ones. The new algorithm is based on the graph and looks for in and out degrees > 1 in the transitive reduction of the graph (meaning, only the minimal edges to keep graph cohesion). A node (spot) can only be linked to one incoming and one outgoing edge, because edges are the link between two frames, and there can only be one spot, hence one link, per frame in a minimal graph. The algorithm only keeps the edge with the node that shares the most neighbors (predecessors + successors) in the non-reduced graph. Indeed, because links are made for each delta frame, spots can have more than one incoming/outgoing edge, although they are redundant. These edges form the neighborhood of each spot, and two spots with similar neighborhood are more likely to be in the same region, and hence track.

Additionally, I created a utils/ subpackage, that contain useful functions, such as color picking (colors.py) and transitive reductions/closures (graph.py). This subpackage can be extended at will to provide more useful functions (notably, those gaussian function that I have in every file or so.)